Create form logic has its own module now.
