+++ src/app/sensors/create-event/create-event.component.ts
@@ -1,14 +1,11 @@
-import {Component, Inject} from '@angular/core';
-import {
-  FormGroup,
-  FormBuilder,
-  Validators,
-  FormArray,
-} from '@angular/forms';
-import {DeviceTypes} from "../../models/deviceType";
-import {SensorService} from "../../service/sensor/sensor.service";
-import {DeviceService} from "../../service/deviceService/device.service";
-import {MAT_DIALOG_DATA, MatDialogRef} from "@angular/material/dialog";
+import { Component, Inject } from '@angular/core';
+import { FormBuilder, FormGroup, FormArray, Validators } from '@angular/forms';
+import { MatDialogRef, MAT_DIALOG_DATA } from "@angular/material/dialog";
+
+import { DeviceTypes } from "../../models/deviceType";
+import { SensorService } from "../../service/sensor/sensor.service";
+import { DeviceService } from "../../service/deviceService/device.service";
+import { EventFormFactory } from "./event-form-factory";
 
 @Component({
   selector: 'app-create-event',
@@ -17,6 +14,8 @@ import {MAT_DIALOG_DATA, MatDialogRef} from "@angular/material/dialog";
 })
 export class CreateEventComponent {
   userId!: string;
+
+  private eventFormFactory: EventFormFactory;
 
   sensors: any[] = [];
   devices: any[] = [];
@@ -35,36 +34,16 @@ export class CreateEventComponent {
     this.userId = data.userId;
 
     this.initDevices().then(r => {
-    })
-    this.addTriggerForm = this.createTriggerForm(this.deviceTypes[0]);
+    });
+
+    this.eventFormFactory = new EventFormFactory();
+    this.eventFormFactory = new EventFormFactory();
+    this.addTriggerForm = this.eventFormFactory.createTriggerForm(this.deviceTypes[0]);
     this.registerDeviceTypeChange(0);
   }
 
   get actions(): FormArray {
     return this.addTriggerForm.get('actions') as FormArray;
-  }
-
-  createTriggerForm(selectedDeviceType: string): FormGroup {
-    return this.formBuilder.group({
-      sensor: ['', Validators.required],
-      name: ['', Validators.required],
-      description: ['', Validators.required],
-      value: [0, [Validators.required, Validators.min(0), Validators.max(1023)]],
-      condition: ['', Validators.required],
-      enabled: ['', Validators.required],
-      resetValue: [0, [Validators.required, Validators.min(0), Validators.max(1023)]],
-      actions: this.formBuilder.array([
-        this.createActionFormGroup(selectedDeviceType),
-      ], Validators.required),
-    });
-  }
-
-  createActionFormGroup(deviceType: string): FormGroup {
-    return this.formBuilder.group({
-      deviceType: ['', Validators.required],
-      deviceId: ['', Validators.required],
-      state: this.getStateFormGroup("null"),
-    });
   }
 
   addNewAction(): void {
@@ -86,7 +65,7 @@ export class CreateEventComponent {
     }
 
     try {
-      await this.sensorService.addTriggerToProfile(this.userId, formValue);
+      await this.sensorService.addEventToProfile(this.userId, formValue);
       this.dialogRef.close();
     } catch (error) {
       console.error('Error saving trigger to the database:', error);
@@ -94,28 +73,11 @@ export class CreateEventComponent {
   }
 
   getStateFormGroup(deviceType: string): FormGroup {
-    const stateFormGroups: { [key: string]: FormGroup } = {
-      toggle: this.formBuilder.group({state: [false, Validators.required]}),
+    return this.eventFormFactory.getStateFormGroup(deviceType);
+  }
 
-      openLock: this.formBuilder.group({
-        locked: [false, Validators.required],
-        open: [false, Validators.required],
-      }),
-
-      fan: this.formBuilder.group({
-        on: [false, Validators.required],
-        reverse: [false, Validators.required],
-      }),
-
-      screen: this.formBuilder.group({
-        on: [false, Validators.required],
-        text: ['', [Validators.required, Validators.maxLength(16)]],
-      }),
-
-      buzzer: this.formBuilder.group({tune: ['alarm', Validators.required]}),
-    };
-
-    return stateFormGroups[deviceType] || this.formBuilder.group({});
+  createActionFormGroup(deviceType: string): FormGroup {
+    return this.eventFormFactory.createActionFormGroup(deviceType);
   }
 
   registerDeviceTypeChange(index: number): void {
@@ -141,5 +103,4 @@ export class CreateEventComponent {
       console.error('Device not found');
     }
   }
-
 }
